add user permissions for deleting listings

+++ App/controllers/ListingController.php
@@ -141,9 +141,16 @@ class ListingController {
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing Not Found');
             return;
+        }
+
+        // Authorization
+        if (Session::get('user')['id'] !== $listing->user_id) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect('/listings/' . $listing->$id);
         }
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
@@ -153,7 +160,6 @@ class ListingController {
 
         redirect('/listings');
     }
-
 
     /**
      * Show the listing edit form
